Renamed all classes to reflect their actual purpose and with consistent naming scheme. Changed all uses of FragmentManager to childFragmentManager. Restructured all fragments to extend a base fragment to allow for easier implementation of common properties. Rewrote the back press handling system to be more reliable and stable. Removed unnecessary files and classes. Started fixing the appbar title update system.
