--- app/src/main/java/com/example/twainz/CustomViewPager.java
@@ -1,27 +1,0 @@
-package com.example.twainz;
-
-import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewPager;
-import android.util.AttributeSet;
-import android.view.MotionEvent;
-
-public class CustomViewPager extends ViewPager {
-
-    public CustomViewPager(@NonNull Context context, @Nullable AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    @Override
-    public boolean onInterceptTouchEvent(MotionEvent ev) {
-//         return super.onInterceptTouchEvent(ev);
-        return false;
-    }
-
-    @Override
-    public boolean onTouchEvent(MotionEvent ev) {
-//        return super.onTouchEvent(ev);
-        return false;
-    }
-}
