+++ app/src/main/java/com/example/twainz/FragmentRoot.java
@@ -25,7 +25,7 @@ public class FragmentRoot extends Fragment implements BackPressListener {
 
             if (!fragmentTitles.containsKey(this.getTag()))
                 fragmentTitles.put(this.getTag(), title);
-            
+
         }
     }
